+++ NEWS.md
@@ -1,5 +1,8 @@
 # incidence2 (development version)
 
+## Bug fixes
+* Fixes bug when input object to incidence is a data.table.
+
 # incidence2 1.2.1
 
 ## Bug fixes